Move out generic job classes
